remoção da sidebar e adicionando segunda etapa para remoção de itens.

+++ src/app/components/cart/cart.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Item } from 'src/app/shared/interfaces/item';
 import { CartService } from 'src/app/shared/services/cart.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cart',
@@ -21,12 +22,24 @@ export class CartComponent implements OnInit {
   }
 
   excluirItemCarrinho(item: Item) {
-    const index = this.itensCarrinho.indexOf(item);
-    if (index !== -1) {
-      this.itensCarrinho.splice(index, 1);
-      this.carrinhoService.excluirItemDoCarrinho(item);
-      this.calcularTotal();
-    }
+    Swal.fire({
+      html: `Você tem certeza que deseja <b>${item.nome}</b> remover do seu carrinho?`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Remover item!',
+      cancelButtonText: 'cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const index = this.itensCarrinho.indexOf(item);
+        if (index !== -1) {
+          this.itensCarrinho.splice(index, 1);
+          this.carrinhoService.excluirItemDoCarrinho(item);
+          this.calcularTotal();
+        }
+      }
+    })
   }
 
   calcularTotal() {
